+++ src/System.Linq.Expressions/src/System.Linq.Expressions.csproj
@@ -160,11 +160,8 @@
   <ItemGroup>
     <Compile Include="System\Linq\Expressions\DynamicExpressionVisitor.cs" />
     <Compile Include="System\Linq\Expressions\DynamicExpression.cs" />
-    <Compile Include="System\Linq\Expressions\Expression.netstandard.cs" />
-    <Compile Include="System\Linq\Expressions\LambdaExpression.netstandard.cs" />
-    <Compile Include="System\Linq\Expressions\ExpressionVisitor.netstandard.cs" />
     <Compile Include="System\Linq\Expressions\Compiler\TypeInfoExtensions.cs" />
-    <Compile Include="System\Linq\Expressions\Compiler\DelegateHelpers.netstandard.cs" />
+    <Compile Include="System\Linq\Expressions\Compiler\DelegateHelpers.cs" />
     <Compile Include="System\Runtime\CompilerServices\RuleCache.cs" />
     <Compile Include="System\Runtime\CompilerServices\CallSite.cs" />
     <Compile Include="System\Runtime\CompilerServices\CallSiteBinder.cs" />
